remove serde_codegen
