--- libs/bragi/build.rs
@@ -1,19 +1,0 @@
-#[cfg(feature = "serde_codegen")]
-fn main() {
-    extern crate serde_codegen;
-
-    use std::env;
-    use std::path::Path;
-
-    let out_dir = env::var_os("OUT_DIR").unwrap();
-
-    let src = Path::new("src/model.in.rs");
-    let dst = Path::new(&out_dir).join("model.rs");
-
-    serde_codegen::expand(&src, &dst).unwrap();
-}
-
-#[cfg(not(feature = "serde_codegen"))]
-fn main() {
-    // do nothing
-}
