+++ libs/bragi/src/lib.rs
@@ -28,7 +28,6 @@
 // https://groups.google.com/d/forum/navitia
 // www.navitia.io
 
-#[cfg(feature = "serde_derive")]
 #[macro_use]
 extern crate serde_derive;
