+++ libs/mimir/Cargo.toml
@@ -2,7 +2,6 @@
 name = "mimir"
 version = "1.2.0"
 authors = ["Canal TP <dt.ro@canaltp.fr>"]
-build = "build.rs"
 
 [dependencies]
 log = "*"
@@ -11,19 +10,10 @@ docopt = "*"
 rustc-serialize = "*"
 rs-es = "0.5"
 serde = "0.8"
-serde_derive = { version = "0.8", optional = true }
+serde_derive = "0.8"
 serde_json = "0.8"
 regex = "*"
 chrono = "0.2"
 hyper = "0.9"
 geo = "^0.3"
-# we temporarily use our own git while geojson hasn't merged our PR on deserialization
 geojson = { version="*", features=["with-serde"]}
-
-
-[features]
-default = ["serde_codegen"]
-nightly = ["serde_derive"]
-
-[build-dependencies]
-serde_codegen = { version = "0.8", optional = true }
